requirements.txt
# Database
SQLAlchemy==2.0.23
psycopg2-binary>=2.9.9  # PostgreSQL driver

# NLP and Text Processing
spacy==3.7.2
sentence-transformers==2.5.1

# Job Scheduling
APScheduler==3.10.4

# Gmail API and Authentication
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Gemini AI
google-genai>=1.0.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# CLI
click==8.1.7

# File Processing
PyPDF2==3.0.1

# HTTP Requests for API Integration
requests==2.31.0
apify-client==1.6.3

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0

# ML Dependencies
scikit-learn>=1.3.0

# Web App (FastAPI + React)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
python-multipart>=0.0.9
importlib-metadata>=6.0.0; python_version < "3.10"
